requirements.txt
flask>=2.0.0
openai>=1.0.0
python-dotenv>=0.19.0
bcrypt>=3.2.0
faiss-cpu>=1.7.0
numpy>=1.20.0
pymupdf>=1.18.0
tqdm>=4.60.0